Python
Imported the package used for CLI argument
passing,made it by default use Serbian or
su due to the method to_latin by default
being in Serbian. Using the flags -l or
--lang you can specify one of the other
languages supported by the package as well
as a way to throw an error in case the
language doesn't exist.